- Avoid code duplications in V004xClient by moving the implementations up to the AbstractClient. - Minor improvements (typos, return types, ...). - Removed old methods that are not required any more.
