+++ templates/about.html
@@ -11,7 +11,7 @@
 <h2>How to get an account?</h2>
 <ul>
     <li>
-        <b>If you are a faculty member</b>, you likely already have an LDAP account. It is usually your u:account with a "cs" suffix. This account is used to login to the cluster.<br>
+        <b>If you are a faculty member</b>, you likely already have an LDAP account. It is usually your u:account with a "cs" suffix. This account is used to log in to the cluster.<br>
         However, you have to be added to the list of "allowed" accounts (i.e., to the slurm database). Please write an email and include the following information, so we can add you:
         <ul>
             <li>Your CS account name.</li>
